+++ tests/browser.test.js
@@ -66,6 +66,7 @@ describe("Browser", function() {
         assert.strictEqual(headerElement.length, 0);
 
     });
+
     it("Find By Text Containing", async() => {
         await browser.open(configUrls.index);
         const elements = await browser.findByTextContaining("paragraph");
@@ -74,5 +75,22 @@ describe("Browser", function() {
         assert.strictEqual(elements[0].tagName, "p");
         assert.strictEqual(elements[1].textContent, "My second paragraph");
         assert.strictEqual(elements[1].tagName, "p");
+    });
+
+    it("Url", async() => {
+        await browser.open(configUrls.index);
+        assert.strictEqual(await browser.url(), "http://localhost:3456/index.html");
+    });
+
+    it("Changing Url", async() => {
+        await browser.open(configUrls.index);
+        assert.strictEqual(await browser.url(), "http://localhost:3456/index.html");
+        await browser.open(configUrls.click);
+        assert.strictEqual(await browser.url(), "http://localhost:3456/click.html");
+    });
+
+    it("Url Before Opening", async() => {
+        const browser2 = await Wendigo.createBrowser();
+        assert.strictEqual(await browser2.url(), null);
     });
 });