+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
   * Added browser.class and browser.assert.class
+  * Added browser.url and browser.assert.url
 
 0.2.0 / 2018-01-27
 ==================